requirements.txt
streamlit
google-generativeai
Pillow
reportlab
PyPDF2